refactor: share generated contract helpers
